FUNCTIONS/requirements.txt
azure-functions
beautifulsoup4==4.9.0
lxml==4.5.0
requests==2.7.0
requests-futures==1.0.0
pyodbc==4.0.30